requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
qdrant-client==1.7.3
sentence-transformers==2.5.1
semantic-text-splitter==0.1.0
requests==2.31.0
beautifulsoup4==4.12.3
jinja2==3.1.3